+++ monkey/monkey_island/cc/services/attack/technique_reports/pba_technique.py
@@ -22,7 +22,7 @@ class PostBreachTechnique(AttackTechnique, metaclass=abc.ABCMeta):
         """
         :param post_breach_action_names: Names of post-breach actions with which the technique is
         associated
-        (example - `["Communicate as new user"]` for T1136)
+        (example - `["Communicate as backdoor user"]` for T1136)
         :return: Mongo query that parses attack telemetries for a simple report component
         (gets machines and post-breach action usage).
         """